refactor: extract installLocalStorageMock + shared episode fixtures
- Add `installLocalStorageMock()` helper to collapse the 5-line
  `Object.defineProperty(window, "localStorage", ...)` boilerplate
  across 4 test files (8 beforeEach blocks) into a single call.
- Move `validEpisode`/`validEpisode2` fixtures to
  `src/test/fixtures/audio-episode.ts` so both lib and action tests
  share one source of truth. Re-export from the existing actions
  `__fixtures.ts` to preserve existing import sites.
